requirements.txt
coverage==4.5.1
Django==2.0.7
djangorestframework==3.8.2
pytz==2018.5
